requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.35.1
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
bcrypt==4.1.2
